Funktionen erledigt: ein besuchtes Fach, in dem noch keine Note hinterlegt wurde, kann gelöscht werden; der Nutzer kann pro Jahr mehrere verschiedene Fächer besuchen
