+++ src/main/java/com/example/gradecalculator/mapper/SubjectMapper.java
@@ -1,11 +1,10 @@
 package com.example.gradecalculator.mapper;
 
 import com.example.gradecalculator.entities.Subject;
+import com.example.gradecalculator.model.SubjectTO;
 import org.mapstruct.Mapper;
 
 @Mapper
 public abstract class SubjectMapper {
-    public static Object toDTO(Subject subject) {
-        return null;
-    }
+    public abstract SubjectTO dataTO(Subject subject);
 }
