+++ src/main/java/com/example/gradecalculator/repository/SubjectRepository.java
@@ -2,15 +2,15 @@ package com.example.gradecalculator.repository;
 
 import com.example.gradecalculator.entities.Subject;
 import jakarta.transaction.Transactional;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
 
-public interface SubjectRepository extends JpaRepository<Subject, Long> {
+public interface SubjectRepository extends CrudRepository<Subject, Long> {
     Subject findById(long id);
     @Modifying
     @Transactional
