+++ src/main/java/com/example/gradecalculator/repository/UserSubjectRepository.java
@@ -1,15 +1,15 @@
 package com.example.gradecalculator.repository;
 
 import com.example.gradecalculator.entities.UserSubject;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.Set;
 
 @Repository
-public interface UserSubjectRepository extends JpaRepository<UserSubject, Long> {
+public interface UserSubjectRepository extends CrudRepository<UserSubject, Long> {
 
     @Query("SELECT us FROM UserSubject us WHERE us.schoolYear.name = :year AND us.user.id = :userId")
     Set<UserSubject> findBySchoolYearAndUserId(@Param("year") int schoolYear, @Param("userId") Long userId);
@@ -19,5 +19,6 @@ public interface UserSubjectRepository extends JpaRepository<UserSubject, Long> 
 
     @Query("SELECT us FROM UserSubject us WHERE us.schoolYear.name = :schoolYearName AND us.user.id = :userId")
     Set<UserSubject> findBySchoolYearNameAndUserId(@Param("schoolYearName") String schoolYearName, @Param("userId") Long userId);}
+
